Be pedantic and use unsigned int here as uwe@ suggested.

+++ usr.bin/make/lst.lib/lstInt.h
@@ -1,4 +1,4 @@
-/*	$NetBSD: lstInt.h,v 1.14 2008/02/05 15:11:50 joerg Exp $	*/
+/*	$NetBSD: lstInt.h,v 1.15 2008/02/05 16:39:26 joerg Exp $	*/
 
 /*
  * Copyright (c) 1988, 1989, 1990, 1993
@@ -47,7 +47,7 @@
 typedef struct ListNode {
 	struct ListNode	*prevPtr;   /* previous element in list */
 	struct ListNode	*nextPtr;   /* next in list */
-	unsigned short	useCount:8, /* Count of functions using the node.
+	unsigned int	useCount:8, /* Count of functions using the node.
 				     * node may not be deleted until count
 				     * goes to 0 */
  	    	    	flags:8;    /* Node status flags */
